+++ oap-server/exporter/src/test/java/org/apache/skywalking/oap/server/exporter/provider/grpc/GRPCExporterTest.java
@@ -23,38 +23,47 @@ import io.grpc.Server;
 import io.grpc.inprocess.InProcessChannelBuilder;
 import io.grpc.inprocess.InProcessServerBuilder;
 import io.grpc.util.MutableHandlerRegistry;
+import java.util.concurrent.TimeUnit;
+import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.metrics.MetricsMetaInfo;
 import org.apache.skywalking.oap.server.core.analysis.metrics.WithMetadata;
 import org.apache.skywalking.oap.server.core.exporter.ExportData;
 import org.apache.skywalking.oap.server.core.exporter.ExportEvent;
 import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
+import org.apache.skywalking.oap.server.exporter.grpc.ExportMetricValue;
+import org.apache.skywalking.oap.server.exporter.grpc.KeyValue;
 import org.apache.skywalking.oap.server.exporter.grpc.MetricExportServiceGrpc;
+import org.apache.skywalking.oap.server.exporter.grpc.SubscriptionMetric;
 import org.apache.skywalking.oap.server.exporter.provider.ExporterSetting;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
 import org.powermock.reflect.Whitebox;
-
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
 
 import static org.apache.skywalking.oap.server.core.exporter.ExportEvent.EventType.INCREMENT;
+import static org.mockito.Mockito.when;
 
 public class GRPCExporterTest {
 
     private GRPCMetricsExporter exporter;
 
     private MetricExportServiceGrpc.MetricExportServiceImplBase service = new MockMetricExportServiceImpl();
-    private MetricsMetaInfo metaInfo = new MetricsMetaInfo("mock-metrics", DefaultScopeDefine.SERVICE);
+    private MetricsMetaInfo metaInfo = new MetricsMetaInfo(
+        "first-mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true));
 
-    private MetricExportServiceGrpc.MetricExportServiceBlockingStub stub;
-
+    private MetricExportServiceGrpc.MetricExportServiceBlockingStub blockingStub;
+    private MetricExportServiceGrpc.MetricExportServiceStub futureStub;
     private Server server;
     private ManagedChannel channel;
     private MutableHandlerRegistry serviceRegistry;
+    private MockedStatic<DefaultScopeDefine> defineMockedStatic;
 
     @BeforeEach
     public void setUp() throws Exception {
@@ -74,10 +83,14 @@ public class GRPCExporterTest {
         setting.setGRPCTargetHost("localhost");
         setting.setGRPCTargetPort(9870);
         exporter = new GRPCMetricsExporter(setting);
-        serviceRegistry.addService(service);
-        stub = MetricExportServiceGrpc.newBlockingStub(channel);
-        Whitebox.setInternalState(exporter, "blockingStub", stub);
         exporter.start();
+        serviceRegistry.addService(service);
+        blockingStub = MetricExportServiceGrpc.newBlockingStub(channel);
+        futureStub = MetricExportServiceGrpc.newStub(channel);
+        Whitebox.setInternalState(exporter, "blockingStub", blockingStub);
+        Whitebox.setInternalState(exporter, "exportServiceFutureStub", futureStub);
+        defineMockedStatic = Mockito.mockStatic(DefaultScopeDefine.class);
+        when(DefaultScopeDefine.inServiceCatalog(1)).thenReturn(true);
     }
 
     @AfterEach
@@ -96,25 +109,38 @@ public class GRPCExporterTest {
             channel = null;
             server.shutdownNow();
             server = null;
+            defineMockedStatic.close();
         }
     }
 
     @Test
     public void export() {
-        ExportEvent event = new ExportEvent(new MockExporterMetrics(), ExportEvent.EventType.TOTAL);
+        exporter.fetchSubscriptionList();
+        ExportEvent event = new ExportEvent(new MockExporterMetrics(), INCREMENT);
         exporter.export(event);
+        List<SubscriptionMetric> subscriptionList = Whitebox.getInternalState(exporter, "subscriptionList");
+        Assertions.assertEquals("mock-metrics", subscriptionList.get(0).getMetricName());
+        Assertions.assertEquals("int-mock-metrics", subscriptionList.get(1).getMetricName());
+        Assertions.assertEquals("long-mock-metrics", subscriptionList.get(2).getMetricName());
+        Assertions.assertEquals("labeled-mock-metrics", subscriptionList.get(3).getMetricName());
     }
 
-    public static class MockExporterMetrics extends MockMetrics implements WithMetadata {
+    public static class MockExporterMetrics extends MockLabeledValueMetrics implements WithMetadata {
         @Override
         public MetricsMetaInfo getMeta() {
-            return new MetricsMetaInfo("mock-metrics", DefaultScopeDefine.SERVICE);
+            return new MetricsMetaInfo(
+                "labeled-mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true));
         }
     }
 
     @Test
     public void initSubscriptionList() {
         exporter.fetchSubscriptionList();
+        List<SubscriptionMetric> subscriptionList = Whitebox.getInternalState(exporter, "subscriptionList");
+        Assertions.assertEquals("mock-metrics", subscriptionList.get(0).getMetricName());
+        Assertions.assertEquals("int-mock-metrics", subscriptionList.get(1).getMetricName());
+        Assertions.assertEquals("long-mock-metrics", subscriptionList.get(2).getMetricName());
+        Assertions.assertEquals("labeled-mock-metrics", subscriptionList.get(3).getMetricName());
     }
 
     @Test
@@ -124,9 +150,14 @@ public class GRPCExporterTest {
 
     @Test
     public void consume() {
-
         exporter.consume(dataList());
         exporter.consume(Collections.emptyList());
+        List<ExportMetricValue> exportMetricValues = ((MockMetricExportServiceImpl) service).exportMetricValues;
+        Assertions.assertEquals(3, exportMetricValues.size());
+        Assertions.assertEquals(12, exportMetricValues.get(0).getMetricValues(0).getLongValue());
+        Assertions.assertEquals(1234567891234563312L, exportMetricValues.get(1).getMetricValues(0).getLongValue());
+        Assertions.assertEquals(1000L, exportMetricValues.get(2).getMetricValues(0).getLongValue());
+        Assertions.assertEquals(KeyValue.newBuilder().setKey("labelName").setValue("labelValue").build(), exportMetricValues.get(2).getMetricValues(0).getLabels(0));
     }
 
     @Test
@@ -143,10 +174,14 @@ public class GRPCExporterTest {
 
     private List<ExportData> dataList() {
         List<ExportData> dataList = new LinkedList<>();
-        dataList.add(new ExportData(metaInfo, new MockMetrics(), INCREMENT));
-        dataList.add(new ExportData(metaInfo, new MockIntValueMetrics(), INCREMENT));
-        dataList.add(new ExportData(metaInfo, new MockLongValueMetrics(), INCREMENT));
-        dataList.add(new ExportData(metaInfo, new MockDoubleValueMetrics(), INCREMENT));
+        dataList.add(new ExportData(new MetricsMetaInfo(
+            "mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true)), new MockMetrics(), INCREMENT));
+        dataList.add(new ExportData(new MetricsMetaInfo(
+            "int-mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true)), new MockIntValueMetrics(), INCREMENT));
+        dataList.add(new ExportData(new MetricsMetaInfo(
+            "long-mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true)), new MockLongValueMetrics(), INCREMENT));
+        dataList.add(new ExportData(new MetricsMetaInfo(
+            "labeled-mock-metrics", DefaultScopeDefine.SERVICE, IDManager.ServiceID.buildId("mock-service", true)), new MockLabeledValueMetrics(), INCREMENT));
         return dataList;
     }
 }